Update and rename Task4.cpp to Task04.cpp
